+++ CHANGELOG.rst
@@ -13,6 +13,13 @@ Documentation
 ~~~~~~~~~~~~~
 - Added more info in ``quick-start.ipynb`` (:pr:`316`) `Ray Bell`_.
 
+Breaking changes
+~~~~~~~~~~~~~~~~
+- Renamed `mae_test` to `halfwidth_ci_test` to make this comparative metric
+  generic. Now, it accepts any of the distance metrics functions except for
+  `mape`. The new function has an additional argument called `metric` which is a
+  string and name of the target distance metric. (:pr:`319`) `Taher Chegini`_.
+
 
 xskillscore v0.0.20 (2021-05-08)
 --------------------------------
@@ -266,4 +273,5 @@ Testing
 .. _`mcsitter`: https://github.com/mcsitter
 .. _`Riley X. Brady`: https://github.com/bradyrx
 .. _`Ray Bell`: https://github.com/raybellwaves
+.. _`Taher Chegini`: https://github.com/cheginit
 .. _`Zachary Blackwood`: https://github.com/blackary